refactor(optimize): put the metadata mapping where its callers are
`targets.ts` had accumulated four unrelated jobs behind a name that
promised one: the containers a save writes, the form's own shape, seeding
the form from the archive, and building the patch. Nothing in the filename
said which direction data flowed.

Each now sits with what uses it. Building the patch had one caller, so it
moved into `apply/buildUpdateActions.ts` — turning form values into update
actions is what that file is for — and its tests now go through
`buildUpdateActions` rather than a helper nothing else called. What several
modules share earns a home: `identifiers/containers.ts` for the containers
and their labels, `identifiers/resolveMetadataFixerFormValues.ts` for the
seeding, and `metadata/formValues.ts` for the form slice's type.

Pure relocation: every function keeps its body, and the suite keeps its
count.
