+++ apps/web/src/books/optimize/apply/buildUpdateActions.ts
@@ -1,34 +1,81 @@
-import type { WebArchiveUpdateAction } from "@oboku/archive-metadata/web"
+import type {
+  ArchiveMetadataIdentifier,
+  WebArchiveUpdateAction,
+} from "@oboku/archive-metadata/web"
+import { isIsbnBearingScheme } from "@prose-reader/archive-reader"
 import type { FileInspection } from "../useFileInspection"
-import {
-  hasWritableMetadataTarget,
-  resolveArchiveMetadataPatchPlan,
-  resolveMetadataFixerFormValues,
-  trimMetadataFixerFormValues,
-} from "../metadata/targets"
+import { resolveMetadataTargets } from "../metadata/identifiers/containers"
+import { resolveMetadataFixerFormValues } from "../metadata/identifiers/resolveMetadataFixerFormValues"
 import {
   hasImageCompressionOperation,
   parseDimension,
   type BookOptimizeFormValues,
 } from "../form"
 
+const normalizeFormIsbn = (isbn: string): string | undefined => {
+  const trimmed = isbn.trim()
+
+  return trimmed === "" ? undefined : trimmed
+}
+
+const announcesIsbn = ({ scheme }: ArchiveMetadataIdentifier): boolean =>
+  isIsbnBearingScheme(scheme)
+
+/**
+ * The identifiers the archive should end up carrying: the ones it already has,
+ * with the edited ISBN in place of every one of them that announced an ISBN.
+ *
+ * More than one usually does — a book whose OPF and ComicInfo both carry it is
+ * read as an `ISBN` and a `GTIN` of the same value — and they are one fact, so
+ * the edit replaces the first and drops the rest. Keeping the others would
+ * leave the previous ISBN behind under the other scheme.
+ *
+ * The patch is the complete set rather than a delta, so every identifier the
+ * book carries has to be listed for it to survive the write. Only the ISBN is
+ * editable here; the rest are passed through as read.
+ */
+const patchIdentifiers = (
+  inspection: FileInspection,
+  isbn: string | undefined,
+): ArchiveMetadataIdentifier[] => {
+  const identifiers = (
+    inspection.resolvedArchive.metadata.identifiers ?? []
+  ).map(function toPatchIdentifier({ scheme, value, unique }) {
+    return { scheme, value, unique: unique === true }
+  })
+  const editedIndex = identifiers.findIndex(announcesIsbn)
+
+  const carried = identifiers.flatMap(
+    function keepOrReplace(identifier, index) {
+      if (!announcesIsbn(identifier)) return [identifier]
+      if (index !== editedIndex || isbn === undefined) return []
+
+      return [{ ...identifier, value: isbn }]
+    },
+  )
+
+  return editedIndex === -1 && isbn !== undefined
+    ? [...carried, { scheme: "ISBN", value: isbn }]
+    : carried
+}
+
 const resolveMetadataPatchAction = (
   values: BookOptimizeFormValues,
   inspection: FileInspection,
 ): WebArchiveUpdateAction | undefined => {
-  if (!hasWritableMetadataTarget(inspection)) return undefined
+  const targets = resolveMetadataTargets(inspection)
 
-  const trimmed = trimMetadataFixerFormValues(values)
-  const resolved = resolveMetadataFixerFormValues(inspection)
+  if (targets.comicInfo !== true && targets.opf !== true) return undefined
 
-  if (trimmed.isbn === resolved.isbn) return undefined
+  const isbn = normalizeFormIsbn(values.isbn)
 
-  const { patch, targets } = resolveArchiveMetadataPatchPlan(
-    trimmed,
-    inspection,
-  )
+  if (isbn === resolveMetadataFixerFormValues(inspection).isbn) return undefined
 
-  return { kind: "patch-metadata", patch, targets }
+  return {
+    kind: "patch-metadata",
+    patch: { identifiers: patchIdentifiers(inspection, isbn) },
+    targets,
+  }
 }
 
 const resolveCompressImagesAction = (
