+++ apps/web/src/books/optimize/metadata/formValues.ts
@@ -1,3 +1,4 @@
+/** The metadata slice of the optimize form. */
 export type MetadataFixerFormValues = {
   isbn: string
 }
