+++ apps/web/src/books/optimize/metadata/identifiers/inspection.fixture.ts
@@ -4,7 +4,7 @@ import {
   resolveArchive,
   type ArchiveRecord,
 } from "@prose-reader/archive-reader"
-import type { FileInspection } from "../useFileInspection"
+import type { FileInspection } from "../../useFileInspection"
 
 const toArrayBuffer = (body: string): ArrayBuffer => {
   const bytes = new TextEncoder().encode(body)
