+++ apps/web/src/books/optimize/metadata/identifiers/resolveMetadataFixerFormValues.ts
@@ -1,0 +1,11 @@
+import { identifierValue } from "@prose-reader/archive-reader"
+import type { FileInspection } from "../../useFileInspection"
+import type { MetadataFixerFormValues } from "../formValues"
+
+export const resolveMetadataFixerFormValues = (
+  inspection: FileInspection,
+): MetadataFixerFormValues => ({
+  isbn:
+    identifierValue(inspection.resolvedArchive.metadata.identifiers, "ISBN") ??
+    "",
+})
